Serialization (#10)
* fix asu_id

* output E-values

* enable model.save

* tidy things up

* i think it's not too broken

* remove legacy api example

* autograph, be happy

* just checking stuff in

* fix reflection subsampling bug

* make normalization happen in the scale module

* laplace why not?

* fixify savification

* monkeypatch better raggedtensor repr

* shhhh

* build method now builds

* make standardardization serialize properly

* use keras for serialization

* make serialization work

* use keras utility for (de)serialization

* sure you can wadam

* basically working for structure factors and wilson

* restored double wilson, not intensity posterior

* resurrect gamma intensity posterior

* refactor data management

* update installer, parstream is merged in rs

* fix reparameterization of rice distribution

* pickles pickles

* yamlify the pickles

* fix shameful bug

* add a performance note

* rice base

* tentatively restore dw prior

* new constructor for racs

* serialize data manager to yml file

* remove incorrect kl div

* restore dense Dkl. work on supporting sparse too

* change sign in reparameterization

* whitespace

* update scaling model

* dunno

* update surrogates, rice_woolfson, cli

* update docstring for reflection subsampling

* build normalization in feedforward layers

* add timeout to online standardization layer

* clean up standardization

* clean up. remove standardization model.

* leaky relus are the default

* biject the location parameter

* remove unused likelihood

---------

+++ abismal/io/stream.py
@@ -72,9 +72,9 @@ class StreamLoader(rs.io.crystfel.StreamLoader):
         I = tf.convert_to_tensor(I[None,:,None])
         d = tf.convert_to_tensor(d[None,:,None])
         SigI = tf.convert_to_tensor(SigI[None,:,None])
-        hkl = tf.convert_to_tensor(hkl[None,:,:], dtype='int32')
+        hkl = tf.convert_to_tensor(hkl[None,:,:], dtype='int64')
         metadata = tf.convert_to_tensor(metadata[None,:,:])
-        asu = tf.ones_like(I, dtype='int32') * self.asu_id
+        asu = tf.ones_like(I, dtype='int64') * self.asu_id
         wavelength = tf.ones_like(I) * wavelength
 
         data = ((asu, hkl, d, wavelength, metadata, I, SigI), (I,))
